refactor: 역할(책임)에 따라 폴더 나누기 파일명 역할에 맞게 나누기(! init파일은 따로 폴더에 둘지, index.js에 그냥 쓰고 사용할지, util폴더에 둘지 고민중 일단 util폴더로)
